Adding a Clock class to the project.
An engine meant for action games inherently requires some sort of time
information to be available to certain parts of the game. For
instance, determining if a player hits an enemy or how to properly
animate elements on the screen. I've thought of essentially two ways
to simulate this in the game: 1) pass a time value from the system
(currently using SFML's built-in Clock class) to the global context
and have it propagate to the objects that need it, or 2) make a
globally-available clock object that is accessible for any given
object to poll.

Currently Evolver uses method 1), probably due to me trying to avoid
some bad design choices I was forced into during the early days of my
OpenGL programming in C with GLUT (dark days indeed). The problem with
this method is that it clutters the interfaces for a number of
objects. For instance, all the of behavior classes attached to Actors
require an "elapsedTime" argument for all methods, a value that's not
always strictly necessary and not conceptually related to moving,
jumping, falling, etc.

Thus the goal of this branch is to implement a Clock object unique to
each instance of the game that can be made available to those objects
in the engine that require timing information and to remove the parts
of the interface that currently rely on timing arguments. By now it
should be obvious that this will be made possible using the Singleton
OOP design pattern. At this point I'm not concerned with making sure
that the engine is capable of multi-threading, so I'm going to
implement the thread-unsafe version of the Singleton pattern. Though
it may be a good idea to prepare the engine for multi-threading, I'll
leave it as an exercise for myself later on.

The Clock object may also include the ability to track the total
amount of time playing the game. Not necessarily a very whiz-bang
feature, but useful nonetheless.