Adding Time class to the project.
The Time class is a convenience class for keeping a simple time
stamp. It's resolution will be set as integral hours and minutes with
a float seconds field. It's most useful functions will be the
overloaded addition ('+') and subtraction ('-') operators.

This will be most useful for the Timer class/interface