Finished implementing TimeStamp class and added a new rule for ignoring Emacs auto save files to .gitignore
I originally intended TimeStamp to hold the seconds value from SFML,
which it stores with seconds as a float with (whole seconds as the
integral part and milliseconds the decimal part), for time-dependent
parts of the game. Given the quick turn-around between executions,
this seemed like a compact way of representing time. The other parts,
hours and minutes, were merely for time stamping save game files,
telling the player total time played, etc.

On reviewing at least one other popular multimedia library (SDL) and
thinking about time representation in Unix-based OSes, time is
generally given in integral "ticks," which represent
milliseconds. Under that assumption, the interface of TimeStamp is
altered to include separate integral values for seconds and
milliseconds instead of using SFML's representation of a single float
value. This will necessitate an adapter function or class to convert
from SFML's float time to standardized integral time, but it will
promote reuse should I ever decide to change multimedia libraries or
use time keeping in other programs. It's (kind of) worth noting that
this will necessitate an extra calculation step for conversion on each
time update, but integer calculations are fast and many, many more are
going to be happening for other reasons.

One small change was also the addition of a rule in .gitignore to
ignore Emacs auto-save files in the working directory.