Clock is modified to use the TimeStamp class.
I originally conceived of the Clock and TimeStamp classes separately,
but on implementing them realized that they had duplicate time
resolution and bookkeeping functionality. Thus Clock is modified to
internally represent time using the TimeStamp class.

Realizing that interfacing with the Clock only through TimeStamp
objects would be awkward for outside classes, Clock uses something of
an adapter pattern to pass time information through to the internal
TimeStamp object. However, should anyone want to take advantage of the
convenience of TimeStamp objects, there are methods for both returning
the Clock's time as a TimeStamp, as well as setting and updating the
time via TimeStamp objects.

I think this is good general-purpose design for not only games, but
any application that needs centralized timekeeping. However, for the
specific purpose of SFML, time information will be coming in as a
float value. This could be taken care of as either an adapter function
that takes the float time value from SFML and passes it millisecond
and second parts separately to the Clock. Alternatively in an
object-oriented manner I could implement an additional Clock, ClockSF,
that extends the Clock and adds a new function to update the time in
terms of a float value and interprets it as the simple adapter
function above would.