Adding files for Timer interface class.
The Timer class will be used for all game objects that rely on time
information. The basic idea of such an interface is that an object
needs to know the elapsed time since some previous event. This will
entail at least having one time marker for the previous event and a
method to get the current time. This of course is possible through the
global Clock that was just created, and conveniently using a TimeStamp
object for the previous event.

For generality's sake, this will likely include functions for getting
the elapsed time at each resolution, as well as one that returns a
TimeStamp representing the full elapsed time.

It may also be worth including an additional TimeStamp for the
original instantiation of the Timer object. This would allow for
calculations based both on the elapsed time since the marked event
(for things like animations) as well the total existence time of the
object for any long-term behaviors.