Changed the '+' and '-' operators for TimeStamp to solve compiler warnings.
This was a warning rather than an error in regards to the fact that
the function returned a locally-created instance of
TimeStamp. However, turning this into a slightly obfuscated one-liner
cleared up the warning. I think it may have something to do with using
the copy constructor rather than the assignment operator for a
pointer. In any case, I don't like the one-liner, but it's better than
the compiler warning.