Removed Observer class compilation and linking from the Makefile.
The Observer class compiles without error (though it is untested), but
it's not currently used in the project (it's eventually going to be
used for Settings objects and possibly in the device
interface). Previously the Makefile attempted to link the compiled
Observer into the final executable, but since it's not used would
produce linker errors related to missing vtable values (or something
like that).