Changing TimeStamp normalization to take care of the case where fields are negative.
Previously TimeStamp only correctly normalized each field as it could
be reduced to a larger one (i.e. 1035 milliseconds would result in an
increment by one to the number of seconds and changing the
milliseconds field to 35). Now normalization also attempts to use an
algorithm akin to elementary school subtraction to "borrow" from a
larger field to bring the smaller field to zero or above (i.e. a
millisecond value of -2023 will be normalized to 77, decrementing
three from the seconds field).

This of course makes the assumption that a TimeStamp has an overall
positive value. That is, when normalized all fields are 0 or
above. Otherwise the normalization will lead to a TimeStamp with an
essentially nonsensical value including a negative "hours" field. This
should only result from stupid values passed to the setter functions
of the TimeStamp or subtracting a later/larger TimeStamp from an
older/smaller one. Both of these situations make little sense and
should not occur if any reason is used by the programmer. However,
this probably indicates that it would be a good idea to change the
setter functions and subtraction operator for TimeStamps to throw
exceptions for illegal inputs.