Changing the behavior of TimeStamp addition and subtraction to not be broken.
Previously the addition and subtraction operators for TimeStamps used
a one-liner trick involving the complex assignment forms (+= and -=)
along with the sopy constructor based on code I had seen
online. Whether my use of pointers differs from those assumed by the
article I was reading or the author was just suggesting crappy code,
the TimeStamp instance returned from the copy constructor was getting
destroyed before its pointer was passed out of the call to the
arithmetic operator. Thus I was getting junk values when using the
arithmetic operators even though the logic involved was overall sound.

Now a less ephemeral instance is created and returned to the caller,
seeming to solve the problem.